+++ packages/cgm_ble/lib/cgm_ble.dart
@@ -1,3 +1,4 @@
 library;
 
+export 'src/ble_failure.dart';
 export 'src/ble_transport.dart';